+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
+## [1.0.1] - 2026-04-06
+
+### Other
+
+- Add Thunderstore icon and packaging
+- Better Thunderstore display name
+- Add LICENSE/CHANGELOG to Thunderstore ZIP, bump manifest version on release
+- Widen changelog artifact paths to catch all release-relevant commits
+
 ## [1.0.0] - 2026-04-06
 
 First release.
-